pyproject.toml
[tool.poetry]
name = "github-analytics"
version = "0.1.0"
description = ""
authors = ["bulatsabirow <bulatsabirow@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["sqlalchemy"], version = "^0.110.0"}
asyncpg = "^0.29.0"
uvicorn = "^0.29.0"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
aiohttp = "^3.9.3"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"
aiosqlite = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
